moved appbar composable to its own file into components package
