+++ app/src/main/java/com/kodeco/android/countryinfo/ui/components/AppBar.kt
@@ -1,0 +1,26 @@
+package com.kodeco.android.countryinfo.ui.components
+
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+
+@Composable
+fun AppBar(title: String, imageVector: ImageVector, iconClickAction: () -> Unit) {
+    TopAppBar(
+        title = {
+            Text(
+                text = title,
+                color = Color.White
+            )
+        },
+        navigationIcon = {
+            IconButton(onClick = { iconClickAction.invoke() }) {
+                Icon(imageVector = imageVector, contentDescription = "")
+            }
+        }
+    )
+}
